refactor(m20): keep mapping in existing files
